# nav-sandhu-files/explorer/app/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Hyperledger Explorer REST API Swagger",
        "description": "Rest API for fabric",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hyperledger Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "https://localhost:8080"
        }
    ],
...
    "paths": {
...
        "/api/status/{channel_genesis_hash}": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to get Count of all Channel details",
                "operationId": "getStatusByHash",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please fill channel genesis hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/chaincode/{channel_genesis_hash}": {
            "get": {
                "tags": ["Chaincodes"],
                "summary": "Query to fetch all Installed/instantiated chaincodes",
                "operationId": "getChaincode",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Query to fetch all Installed/instantiated chaincodes",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/block/transactions/{channel_genesis_hash}/{number}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to Fetch Block Transactions",
                "operationId": "getBlockTransactions",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Please give the number value",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blockActivity/{channel_genesis_hash}": {
            "get": {
                "tags": ["Blocks"],
                "summary": "Query to fetch BlockActivity List",
                "operationId": "getBlockActivity",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Query to fetch BlockActivity List",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/transaction/{channel_genesis_hash}/{txid}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch transactions by txId",
                "operationId": "getTransactionById",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "txid",
                        "in": "path",
                        "description": "Please provide transaction id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/txList/{channel_genesis_hash}/{number}/{txid}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to get transaction List",
                "operationId": "getTransactionList",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Please give the number Value",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "txid",
                        "in": "path",
                        "description": "Please give the txid value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/peers/{channel_genesis_hash}": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to fetch all Installed/instantiated chaincodes",
                "operationId": "getPeers",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Query to fetch all Installed/instantiated chaincodes",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blockAndTxList/{channel_genesis_hash}/{number}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch blocks and transaction list",
                "operationId": "getBlockAndTxList",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Please give the number value",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/txByMinute/{channel_genesis_hash}/{hours}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch transaction by minute list",
                "operationId": "getTransactionByMin",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hours",
                        "in": "path",
                        "description": "Please give the hour value",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/txByHour/{channel_genesis_hash}/{days}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch blocks and transaction list",
                "operationId": "getTxByHour",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "days",
                        "in": "path",
                        "description": "Please give the day value",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}